requirements.txt
Django==2.2.24
drf-yasg==1.20.0
djangorestframework==3.11.2
typing-extensions==3.7.4.3